refactor: decompose repl.py into smaller modules
- Created repl_components package with modular components
- Extracted command_parser.py for argument parsing (35 lines)
- Extracted tool_executor.py for tool handling (85 lines)
- Extracted output_display.py for output formatting (34 lines)
- Extracted error_recovery.py for error recovery (89 lines)
- Reduced repl.py from 578 to 391 lines
- Updated all tests to work with new module structure
- All characterization tests passing
