+++ src/tunacode/cli/repl_components/__init__.py
@@ -1,0 +1,10 @@
+"""
+REPL components package for modular REPL functionality.
+"""
+
+from .command_parser import parse_args
+from .error_recovery import attempt_tool_recovery
+from .output_display import display_agent_output
+from .tool_executor import tool_handler
+
+__all__ = ["parse_args", "attempt_tool_recovery", "display_agent_output", "tool_handler"]
