+++ src/tunacode/cli/repl_components/command_parser.py
@@ -1,0 +1,34 @@
+"""
+Module: tunacode.cli.repl_components.command_parser
+
+Command parsing utilities for the REPL.
+"""
+
+import json
+
+from tunacode.exceptions import ValidationError
+from tunacode.types import ToolArgs
+
+
+def parse_args(args) -> ToolArgs:
+    """
+    Parse tool arguments from a JSON string or dictionary.
+
+    Args:
+        args (str or dict): A JSON-formatted string or a dictionary containing tool arguments.
+
+    Returns:
+        dict: The parsed arguments.
+
+    Raises:
+        ValueError: If 'args' is not a string or dictionary, or if the string is not valid JSON.
+    """
+    if isinstance(args, str):
+        try:
+            return json.loads(args)
+        except json.JSONDecodeError:
+            raise ValidationError(f"Invalid JSON: {args}")
+    elif isinstance(args, dict):
+        return args
+    else:
+        raise ValidationError(f"Invalid args type: {type(args)}")
